+++ src/sad_spirit/pg_wrapper/converters/FloatConverter.php
@@ -20,51 +20,14 @@ declare(strict_types=1);
 
 namespace sad_spirit\pg_wrapper\converters;
 
-use sad_spirit\pg_wrapper\exceptions\TypeConversionException;
-
 /**
  * Converter for float type, handles corner cases like NaN and Infinity
  */
-class FloatConverter extends BaseConverter
+class FloatConverter extends NumericConverter
 {
     protected function inputNotNull(string $native)
     {
-        $native = trim($native);
-        if (is_numeric($native)) {
-            return (double)$native;
-
-        } elseif (0 === strcasecmp($native, 'NaN')) {
-            return NAN;
-
-        } elseif (0 === strcasecmp($native, 'Infinity')) {
-            return INF;
-
-        } elseif (0 === strcasecmp($native, '-Infinity')) {
-            return -INF;
-
-        } else {
-            throw TypeConversionException::unexpectedValue($this, 'input', 'numeric value', $native);
-        }
-    }
-
-    protected function outputNotNull($value): string
-    {
-        if (is_float($value)) {
-            if (is_nan($value)) {
-                return 'NaN';
-            } elseif (is_infinite($value)) {
-                return $value > 0 ? 'Infinity' : '-Infinity';
-            } else {
-                return strtr((string)$value, ',', '.');
-            }
-        }
-        if (is_string($value)) {
-            $value = strtr($value, ',', '.');
-        }
-
-        if (!is_scalar($value) || !is_numeric($value)) {
-            throw TypeConversionException::unexpectedValue($this, 'output', 'numeric value', $value);
-        }
-        return (string)$value;
+        $result = parent::inputNotNull($native);
+        return is_string($result) ? (double)$result : $result;
     }
 }